+++ examples/rustapi_module/src/subclassing.rs
@@ -13,6 +13,13 @@ impl Subclassable {
     }
 }
 
+#[pyproto]
+impl pyo3::PyObjectProtocol for Subclassable {
+    fn __str__(&self) -> PyResult<&'static str> {
+        Ok("Subclassable")
+    }
+}
+
 #[pymodule]
 fn subclassing(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_class::<Subclassable>()?;