+++ examples/rustapi_module/tests/test_subclassing.py
@@ -6,10 +6,11 @@ PYPY = platform.python_implementation() == "PyPy"
 
 
 class SomeSubClass(Subclassable):
-    pass
+    def __str__(self):
+        return "Subclass"
 
 
 def test_subclassing():
     if not PYPY:
         a = SomeSubClass()
-        _b = str(a) + repr(a)
+        assert str(a) == "Subclass"